+++ Source/Core/Render/Graphics/IGeometryModifier.h
@@ -10,7 +10,7 @@ struct D3D11_MAPPED_SUBRESOURCE;
 class IGeometryModifier
 {
 public:
-	virtual void SetRaw(const char* name, const void* data, const size_t byteSize) = 0;
+	virtual void SetRaw(const char* name, const void* data, const uint32_t byteSize) = 0;
 	virtual void SetFloat(const char* name, const float& value) = 0;
 	virtual void SetFloat3(const char* name, const float* value) = 0;
 	virtual void SetMatrix(const char* name, const XMFLOAT4X4& value) = 0;
